refactor: polishing
